Don't check in generated files
This was originally done for convenience, so that the
res_moodle_stack didn't need to build the plugin. But it makes
it difficult to keep track of what's going on, and means this
repo has lots of cruft in it and contradicts common sense.

Note that removing these files will make res_moodle_stack more
complex, as it will have to build the repository plugin itself.